requirements.txt
requests
openai==0.28.1
google-api-python-client==2.131.0
google-auth==2.33.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
APScheduler==3.10.4
pytz==2024.1
python-dotenv==1.0.1
